Introduce audit transactions.
This PR introduces "audit" as a new, distinct transaction type.  An audit
transaction expresses a constraint on a user's balance or asset balance.

Besides the payment of the sender to the fee sink, an audit transaction does not
modify the state of the system.  Instead, an audit transaction is valid if and
only if the constraint it specifies is true at the time of confirmation (more
specifically, on the intermediate state determined by its position in a block's
transaction stream).

Audit transactions support three basic ordering constraints against a 64-bit
constant: equality, less-than, and greater-than.  These constraints may be
specified with respect to an account's balance or the balance of one of its
assets.

The most basic use of audit transactions is for the ledger verifiers to attest
to a statement proving an account's solvency (or lack thereof).  Since
transactions are transitively signed by a block's certificate, audit
transactions allow accounts to prove solvency to third-party applications.

A more general use of audit transactions is to specify preconditions and
postconditions to be met before or after a group transaction.

One example of such a constraint is in a TEAL account that allows a user to
issue periodic withdrawals from the account at some rate.  When the account has
insufficient funds for a full withdrawal, it may be useful to allow the user to
empty the account via a close transaction.  An audit transaction allows the
escrow account to require that its balance lie underneath a given threshold for
the transaction to execute.

Another example of a useful audit constraint is in a TEAL escrow account that
implements a collateralized debt.  Debt payments may be recorded by the issuance
of tokens by the escrow account.  If the debtor misses an excessive number of
payments, the creditor may close the account to claim the collateral.  An audit
transaction allows the TEAL script to verify that the creditor has missed some
number of payments.