Give upgrades a flexible delay between acceptance and execution.
Proposers may now specify the delay (in rounds) between accepting a protocol and
executing it.  A flexible delay allows certain upgrades to the protocol (e.g.,
relating to security) to execute faster than others (e.g., relating to new
features).

This delay must fall between some minimum and maximum value.  This change sets
minimum and maximum delay values to be 10,000 and 150,000 rounds, respectively.

Note that this change does not modify the number of votes needed to approve a
upgrade in the first place.

A reference Go implementation may be found here:
https://github.com/algorand/go-algorand/commit/7a415e4855558e66b55785b8342f4e59cb60a2ec