Add a rekeying feature
The feature gives balance records a "spending key" field, gives transactions an optional "RekeyTo" address, and gives SignedTxns an optional "authorization address" field. When a transaction rekeys an account, this updates the account's spending key, and future transactions from this account must have their authorization address match (and thus their signatures verify under) this new spending key. This allows rotating spending keys for security while continuing to receive funds at the same address. It also allows changing a normal account into a multisig account (or changing the membership of a multisig account) without changing its address.

This commit also expands the section on transaction signatures to clarify that signatures are not part of transactions (and thus don't influence the TXID) but rather appear next to transactions in structs called SignedTxns, which is the existing behavior.